Movimentação da Barra Lateral para AfterLoginActivity
-Versão do App 0.4.0 -> 0.4.1;

-Activity AfterLoginActivity recriada para receber a barra lateral de MapsEventsActivity;

-Barra Lateral retirada de MapsEventsActivity e movida para AfterLoginActivity;

-Adicionado icone no FloatingActionButton em MapsEventsActivity.
