+++ app/src/main/AndroidManifest.xml
@@ -26,7 +26,6 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".AfterLoginActivity" />
         <activity android:name=".ChallengeActivity" />
         <activity android:name=".SupportiveFriendActivity" />
         <activity
@@ -49,7 +48,11 @@
             android:name=".MapsFragment"
             android:label="@string/title_activity_maps" />
         <activity android:name=".RegisterEventActivity" />
-        <activity android:name=".RegisterActivity"></activity>
+        <activity android:name=".RegisterActivity" />
+        <activity
+            android:name=".AfterLoginActivity"
+            android:label="@string/title_activity_after_login"
+            android:theme="@style/AppTheme.NoActionBar"></activity>
     </application>
 
 </manifest>
