+++ app/src/main/res/layout/activity_after_login.xml
@@ -1,35 +1,25 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<android.support.v4.widget.DrawerLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/drawer_layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:orientation="vertical"
-    tools:context=".AfterLoginActivity">
+    android:fitsSystemWindows="true"
+    tools:openDrawer="start">
 
-    <!-- Button MapsEvents -->
-    <Button
-        android:id="@+id/btMapsEvents"
+    <include
+        layout="@layout/app_bar_after_login"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent" />
+
+    <android.support.design.widget.NavigationView
+        android:id="@+id/nav_view"
         android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:onClick="btMapsEvents"
-        android:hint="@string/MapsEvents"/>
+        android:layout_height="match_parent"
+        android:layout_gravity="start"
+        android:fitsSystemWindows="true"
+        app:headerLayout="@layout/nav_header_after_login"
+        app:menu="@menu/activity_after_login_drawer" />
 
-    <!-- Button Challenge -->
-    <Button
-        android:id="@+id/btChallenge"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/btMapsEvents"
-        android:onClick="btChallenge"
-        android:hint="@string/challenge"/>
-
-    <!-- Button SupportiveFriend -->
-    <Button
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/btChallenge"
-        android:onClick="btSupportiveFriend"
-        android:hint="@string/SupportiveFriend"/>
-
-</RelativeLayout>
+</android.support.v4.widget.DrawerLayout>
