+++ app/src/main/res/layout/activity_maps_events.xml
@@ -13,13 +13,4 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
 
-    <android.support.design.widget.NavigationView
-        android:id="@+id/nav_view"
-        android:layout_width="wrap_content"
-        android:layout_height="match_parent"
-        android:layout_gravity="start"
-        android:fitsSystemWindows="true"
-        app:headerLayout="@layout/nav_header_maps_events"
-        app:menu="@menu/activity_maps_events_drawer" />
-
 </android.support.v4.widget.DrawerLayout>
