+++ app/src/main/res/layout/content_maps_events.xml
@@ -24,8 +24,10 @@
         android:layout_marginTop="69dp"
         android:name="br.com.garrav.projetogarrav.EventIteractorFragment"/>
 
+    <!-- Floating Action Button -->
     <android.support.design.widget.FloatingActionButton
         android:id="@+id/fabRegisterNewEvent"
+        android:src="@drawable/ic_add_location_black_24dp"
         android:layout_width="58dp"
         android:layout_height="61dp"
         android:layout_alignParentEnd="true"
