+++ app/src/main/res/menu/activity_after_login_drawer.xml
@@ -24,7 +24,6 @@
         <item
             android:id="@+id/nav_logout"
             android:title="@string/logout"/>
-        <!-- Itens padrões do Android Studio -->
         <item
             android:id="@+id/nav_camera"
             android:icon="@drawable/ic_menu_camera"
