+++ app/src/main/res/values/strings.xml
@@ -30,4 +30,5 @@
     <string name="report_problem">Reportar Problema</string>
     <string name="about">Sobre</string>
     <string name="logout">Sair</string>
+    <string name="title_activity_after_login">AfterLoginActivity</string>
 </resources>
